improving version handling and allowing install
