+++ autohelm/meta.py
@@ -1,0 +1,3 @@
+
+__version__ = '0.1.0'
+__author__ = 'ReactiveOps, Inc.'
